gui/requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.20.1
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pefile==2022.5.30
Pillow==9.1.1
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyparsing==3.0.9
pywin32-ctypes==0.2.0
requests==2.28.0
urllib3==1.26.9
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0